+++ src/main/java/recipe/foodbar/usecase/recipe/manager/sort/RatingComparator.java
@@ -1,7 +1,5 @@
 package recipe.foodbar.usecase.recipe.manager.sort;
 
-import recipe.foodbar.entities.Recipe.Recipe;
-
 /**
  * This comparator class is responsible for comparing two recipes and returning an ordering based on rating
  * (used to reorder by rating)
